requirements.txt
numpy
scipy
scikit-learn
hydra-core
omegaconf
tqdm
matplotlib
h5py
ruff
